Generators for math blocks
