+++ generators/generators.js
@@ -124,34 +124,6 @@ export function defineGenerators() {
     return [`keyPressed("${key}")`, javascriptGenerator.ORDER_NONE];
   };
 
-  javascriptGenerator.forBlock["random_seeded_int"] = function (block) {
-    const value_from = getFieldValue(block, "FROM", 0);
-    const value_to = getFieldValue(block, "TO", 10);
-    const value_seed = getFieldValue(block, "SEED", 123456);
-
-    const code = `seededRandom(${value_from}, ${value_to}, ${value_seed})`;
-
-    return [code, javascriptGenerator.ORDER_NONE];
-  };
-
-  javascriptGenerator.forBlock["to_number"] = function (block) {
-    const string = javascriptGenerator.valueToCode(
-      block,
-      "STRING",
-      javascriptGenerator.ORDER_ATOMIC,
-    );
-    const conversionType = block.getFieldValue("TYPE");
-
-    let code;
-    if (conversionType === "INT") {
-      code = `parseInt(${string})`;
-    } else {
-      code = `parseFloat(${string})`;
-    }
-
-    return [code, javascriptGenerator.ORDER_NONE];
-  };
-
   javascriptGenerator.forBlock["procedures_defnoreturn"] = function (block) {
     const functionName = javascriptGenerator.nameDB_.getName(
       block.getFieldValue("NAME"),
@@ -706,23 +678,6 @@ javascriptGenerator.forBlock["xyz"] = function (block) {
   // Generate a tuple representing the vector
   const code = `[${x}, ${y}, ${z}]`;
   return [code, javascriptGenerator.ORDER_ATOMIC];
-};
-
-javascriptGenerator.forBlock["math_random_int"] = function (block) {
-  const from =
-    javascriptGenerator.valueToCode(
-      block,
-      "FROM",
-      javascriptGenerator.ORDER_NONE,
-    ) || "0";
-  const to =
-    javascriptGenerator.valueToCode(
-      block,
-      "TO",
-      javascriptGenerator.ORDER_NONE,
-    ) || "0";
-  const code = "randomInteger(" + from + ", " + to + ")";
-  return [code, javascriptGenerator.ORDER_FUNCTION_CALL];
 };
 
 javascriptGenerator.forBlock["keyword"] = function (block) {
